Character Select Fix Attempt
When testing the game, it successfully switches between game screens but has issues once the game actually starts.  When starting the game, neither of the playable characters is in the Game Scene.  I'm not sure why both players appear to be deleted after selecting one, but it's been confusing to try to fix.